chore(linux): port dsp and audiopassthru for linux compatibility Replace `long` with `int32_t` across all DSP and audiopassthru sources; `long` is 64-bit on linux x86_64 but 32-bit on windows, breaking fixed-layout DSP memory banks. Wrap windows-only APIs (COM init, dbghelp, NvOptimus exports) in `#ifdef _WIN32`; add linux crash handler via execinfo backtrace. Add portability shims in AudioPassthru.h (IMMDevice, WCHAR) for non-windows builds. Fix null-init pointer and swprintf bounds in DfxDspPreset.cpp; force headphone_on_ = IS_FALSE to prevent muffled output on speakers.

+++ dsp/ptComSftDfx/Comsftwr.c
@@ -66,7 +66,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * Note that the hrdwrWriteParameterIfNotBusy uses a hard wait loop to
  * complete the second write operation (the actual value).
  */
-int COMSFTWR_DECL comSftwrWriteParam(PT_HANDLE *hp_comSftwr, long l_offset, long l_val)
+int COMSFTWR_DECL comSftwrWriteParam(PT_HANDLE *hp_comSftwr, int32_t l_offset, int32_t l_val)
 {
 	float *flt_ptr;
 	struct comSftwrHdlType *cast_handle;
@@ -106,7 +106,7 @@ int COMSFTWR_DECL comSftwrWriteParam(PT_HANDLE *hp_comSftwr, long l_offset, long
  *  Currently only processes stereo to stereo or mono to mono signals.
  *
  */
-int COMSFTWR_DECL comSftwrProcessWaveBuffer(PT_HANDLE *hp_comSftwr, long *lp_data, long l_length, 
+int COMSFTWR_DECL comSftwrProcessWaveBuffer(PT_HANDLE *hp_comSftwr, int32_t *lp_data, int32_t l_length, 
                          int i_stereo_in_mode, int i_stereo_out_mode, 
 								 int i_buffer_type)
 {
@@ -187,7 +187,7 @@ int COMSFTWR_DECL comSftwrProcessWaveBuffer(PT_HANDLE *hp_comSftwr, long *lp_dat
 			/* On off toggling for demo */
 			static int count = 0;
 			static int on = 0;
-			long int tmp_buf[16384];
+			int32_t int tmp_buf[16384];
 
 			count += l_length;
 			if( count > 44100 * COMSFTWR_AUTO_DEMO_SECS )
@@ -244,16 +244,16 @@ int COMSFTWR_DECL comSftwrProcessWaveBuffer(PT_HANDLE *hp_comSftwr, long *lp_dat
 		float factor = (float)PC_24BIT_FLOAT_PLUS_CLIP/(float)l_length;
 
 		cast_handle->ComSftwrMeterData.left_in = 
-			(long)( *(float *)&(ComSftwrMeterData[i_processor_index].left_in) * factor); 
+			(int32_t)( *(float *)&(ComSftwrMeterData[i_processor_index].left_in) * factor); 
    
 		cast_handle->ComSftwrMeterData.right_in = 
-			(long)( *(float *)&(ComSftwrMeterData.right_in) * factor); 
+			(int32_t)( *(float *)&(ComSftwrMeterData.right_in) * factor); 
 
 		cast_handle->ComSftwrMeterData[i_processor_index].left_out = 
-			(long)( *(float *)&(ComSftwrMeterData.left_out) * factor); 
+			(int32_t)( *(float *)&(ComSftwrMeterData.left_out) * factor); 
 
 		cast_handle->ComSftwrMeterData.right_out = 
-			(long)( *(float *)&(ComSftwrMeterData.right_out) * factor); 
+			(int32_t)( *(float *)&(ComSftwrMeterData.right_out) * factor); 
 	}
 #endif
 
@@ -275,7 +275,7 @@ int COMSFTWR_DECL comSftwrProcessWaveBuffer(PT_HANDLE *hp_comSftwr, long *lp_dat
  *  Process the passed Active Movie buffer, and send it back (no A/D or D/A).
  *
  */
-int COMSFTWR_DECL comSftwrProcessActiveBuffer(PT_HANDLE *hp_comSftwr, short *sp_data, long l_length, 
+int COMSFTWR_DECL comSftwrProcessActiveBuffer(PT_HANDLE *hp_comSftwr, short *sp_data, int32_t l_length, 
                          int i_stereo_in_mode, int i_stereo_out_mode,
 								 int i_buffer_type)
 {
@@ -289,7 +289,7 @@ int COMSFTWR_DECL comSftwrProcessActiveBuffer(PT_HANDLE *hp_comSftwr, short *sp_
    /* l_length comes in with the buffer size in samples */
 
    /* For now, call original saw style processing */
-   if( comSftwrProcessWaveBuffer(hp_comSftwr, (long *)sp_data, l_length, 
+   if( comSftwrProcessWaveBuffer(hp_comSftwr, (int32_t *)sp_data, l_length, 
                          i_stereo_in_mode, i_stereo_out_mode, i_buffer_type) != OKAY)
 		return(NOT_OKAY);
 
@@ -311,7 +311,7 @@ int COMSFTWR_DECL comSftwrInitDspAlgorithm(PT_HANDLE *hp_comSftwr, realtype r_sa
 #else
 	long long perf_count;
 #endif
-	long sample_count_init;
+	int32_t sample_count_init;
 	struct comSftwrHdlType *cast_handle;
 
 	cast_handle = (struct comSftwrHdlType *)hp_comSftwr;
@@ -330,9 +330,9 @@ int COMSFTWR_DECL comSftwrInitDspAlgorithm(PT_HANDLE *hp_comSftwr, realtype r_sa
 #endif
 
 #ifdef WIN32
-	sample_count_init = (long)(perf_count % (__int64)COMSFTWR_DEMO_SAMPLES_ALLOWED);
+	sample_count_init = (int32_t)(perf_count % (__int64)COMSFTWR_DEMO_SAMPLES_ALLOWED);
 #else
-	sample_count_init = (long)(perf_count % (long long)COMSFTWR_DEMO_SAMPLES_ALLOWED);
+	sample_count_init = (int32_t)(perf_count % (long long)COMSFTWR_DEMO_SAMPLES_ALLOWED);
 #endif
 	cast_handle->sample_count = sample_count_init;
 
@@ -445,7 +445,7 @@ int COMSFTWR_DECL comSftwrSetFunctionIndex(PT_HANDLE *hp_comSftwr, char *cp_dspn
 										  short s_bit_width, int i_use_old_bit_width)
 {
 	int index = 0;
-	long memsize = 0;
+	int32_t memsize = 0;
 	int offset = 0;
 	struct comSftwrHdlType *cast_handle;
 
@@ -826,8 +826,8 @@ int COMSFTWR_DECL comSftwrSetFunctionIndex(PT_HANDLE *hp_comSftwr, char *cp_dspn
  */
 int COMSFTWR_DECL comSftwrAllocDspMem(PT_HANDLE *hp_comSftwr)
 {
-	long memsize_required;
-	long current_memsize;
+	int32_t memsize_required;
+	int32_t current_memsize;
 	float *mem_ptr;
 	struct comSftwrHdlType *cast_handle;
